locales/messages/requirements.txt
PyYAML>=5.4
toml>=0.10
